+++ project.flutter/client_app/ios/Runner/AppDelegate.swift
@@ -12,5 +12,6 @@ import UIKit
 
   func didInitializeImplicitFlutterEngine(_ engineBridge: FlutterImplicitEngineBridge) {
     GeneratedPluginRegistrant.register(with: engineBridge.pluginRegistry)
+    ImportBridge.shared.attach(messenger: engineBridge.applicationRegistrar.messenger())
   }
 }