+++ project.flutter/client_app/linux/flutter/generated_plugins.cmake
@@ -3,7 +3,6 @@
 #
 
 list(APPEND FLUTTER_PLUGIN_LIST
-  gtk
   sw_rend
   url_launcher_linux
   zstandard_linux