+++ project.flutter/client_app/macos/Flutter/GeneratedPluginRegistrant.swift
@@ -5,7 +5,6 @@
 import FlutterMacOS
 import Foundation
 
-import app_links
 import file_picker
 import google_sign_in_ios
 import package_info_plus
@@ -16,7 +15,6 @@ import webview_flutter_wkwebview
 import zstandard_macos
 
 func RegisterGeneratedPlugins(registry: FlutterPluginRegistry) {
-  AppLinksMacosPlugin.register(with: registry.registrar(forPlugin: "AppLinksMacosPlugin"))
   FilePickerPlugin.register(with: registry.registrar(forPlugin: "FilePickerPlugin"))
   FLTGoogleSignInPlugin.register(with: registry.registrar(forPlugin: "FLTGoogleSignInPlugin"))
   FPPPackageInfoPlusPlugin.register(with: registry.registrar(forPlugin: "FPPPackageInfoPlusPlugin"))