+++ project.flutter/client_app/windows/flutter/generated_plugins.cmake
@@ -3,7 +3,6 @@
 #
 
 list(APPEND FLUTTER_PLUGIN_LIST
-  app_links
   share_plus
   sw_rend
   url_launcher_windows